refactor(operator): split semantic_graph relationship builders into a submodule
